+++ mods/CreateDragonPlus/src/main/java/plus/dragons/createdragonsplus/mixin/garnished/FanProcessingTypeMixinForGarnished.java
@@ -1,0 +1,28 @@
+package plus.dragons.createdragonsplus.mixin.garnished;
+
+import com.llamalad7.mixinextras.injector.wrapoperation.Operation;
+import com.llamalad7.mixinextras.injector.wrapoperation.WrapOperation;
+import com.simibubi.create.content.kinetics.fan.processing.FanProcessingType;
+import me.fallenbreath.conditionalmixin.api.annotation.Condition;
+import me.fallenbreath.conditionalmixin.api.annotation.Restriction;
+import net.minecraft.core.BlockPos;
+import net.minecraft.world.level.Level;
+import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.injection.At;
+import plus.dragons.createdragonsplus.common.kinetics.fan.freezing.FreezingFanProcessingType;
+import plus.dragons.createdragonsplus.config.CDPConfig;
+import plus.dragons.createdragonsplus.integration.ModIntegration;
+
+@Restriction(require = @Condition(ModIntegration.Constants.CREATE_GARNISHED))
+@Mixin(FanProcessingType.class)
+public interface FanProcessingTypeMixinForGarnished {
+    @WrapOperation(
+            method = "getAt",
+            at = @At(value = "INVOKE", target = "Lcom/simibubi/create/content/kinetics/fan/processing/FanProcessingType;isValidAt(Lnet/minecraft/world/level/Level;Lnet/minecraft/core/BlockPos;)Z")
+    )
+    private static boolean ignoreDisabledType(FanProcessingType instance, Level level, BlockPos blockPos, Operation<Boolean> original) {
+        if (instance instanceof FreezingFanProcessingType && CDPConfig.recipes().enableBulkFreezing.get())
+           return false;
+        return original.call(instance, level, blockPos);
+    }
+}